fix scaladoc and javadoc generation
